Fixed compile time warning.
git-svn-id: https://svn.code.sf.net/p/osp-toolkit/code/main@50 43700839-f8e2-0310-9689-8ed4146226e7

+++ src/ospsecurity.c
@@ -67,7 +67,7 @@ OSPPSecGetLocalCertInfo(
     OSPTASN1OBJECT **ospvLocalCertInfo);
 
 
-void OpenSSLErrorLog(long errorcode);
+void OpenSSLErrorLog(int errorcode);
 
 /**************************************/
 /* IMPLEMENTATION OF MEMBER FUNCTIONS */
@@ -1446,7 +1446,7 @@ OSPPSecGetSignerCertSubjectName(
     return errorcode;
 }
 
-void OpenSSLErrorLog(long errorcode)
+void OpenSSLErrorLog(int errorcode)
 {
 	BIO		*mem_bio	= OSPC_OSNULL;
 	char	*p				= OSPC_OSNULL;
